common: Fix bugs with submitting DMA BDs.
Rx BDs are constructed only after knowing the size of the receive
transaction, rather than some fixed size, which will fail if the receive
transaction is not a multiple of that fixed size. Tx BDs and Rx BDs now
also support transaction sizes that are not a multiple of the max packet
length.