needwun: Initial commit.
Note: in order for the baremetal code to work, you must set MAX_PKT_LEN
to be no greater than 256 (0x100). There are issues that result in
erratic incorrect behavior that leads to failed tests if the packet size
grows beyond this. I suspect it is a coherency issue but I have not yet
had time to look.